+++ run-scripts/train_script_function.sh
@@ -1,12 +1,12 @@
 #!/bin/bash
-#SBATCH --time=07:30:00
+#SBATCH --time=100:00:00
 #SBATCH --mem=1G
 #SBATCH --output=outputs/%A/%a_output.out
 #SBATCH --error=outputs/%A/%a_error.out
 #SBATCH --mail-type=ALL
 #SBATCH --mail-user=muhammad.sattar@aalto.fi
 #SBATCH --job-name=functional_rew
-#SBATCH --array=0
+#SBATCH --array=0-2
 
 module restore FM3-MicRo
 source activate FM3-MicRo
@@ -15,13 +15,13 @@ source activate FM3-MicRo
 declare -A PARAMS
 
 # Parameter set 0
-PARAMS[0]="--batch-size 64 --exc train --goal-reset False --num-eval 100 --reward-type delta_r --rollout-steps 2048 --text-verbosity True --total-timesteps 2500000 --train-episode-time-limit 2.5 --train-fps None --train-render-fps None --train-verbosity True"
+PARAMS[0]="--batch-size 64 --exc train --goal-reset False --num-eval 100 --reward-type delta_r --rollout-steps 2048 --text-verbosity True --total-timesteps 1000000 --train-episode-time-limit 360 --train-fps 10 --train-render-fps 120 --train-verbosity True"
 
 # Parameter set 1
-PARAMS[1]="--batch-size 64 --exc train --goal-reset False --num-eval 100 --reward-type euclidean --rollout-steps 2048 --text-verbosity True --total-timesteps 2500000 --train-episode-time-limit 15 --train-fps 10 --train-render-fps 120 --train-verbosity True"
+PARAMS[1]="--batch-size 64 --exc train --goal-reset False --num-eval 100 --reward-type euclidean --rollout-steps 2048 --text-verbosity True --total-timesteps 1000000 --train-episode-time-limit 360 --train-fps 10 --train-render-fps 120 --train-verbosity True"
 
 # Parameter set 2
-PARAMS[2]="--batch-size 64 --exc train --goal-reset False --num-eval 100 --reward-type sparse --rollout-steps 2048 --text-verbosity True --total-timesteps 2500000 --train-episode-time-limit 15 --train-fps 10 --train-render-fps 120 --train-verbosity True"
+PARAMS[2]="--batch-size 64 --exc train --goal-reset False --num-eval 100 --reward-type sparse --rollout-steps 2048 --text-verbosity True --total-timesteps 1000000 --train-episode-time-limit 360 --train-fps 10 --train-render-fps 120 --train-verbosity True"
 
 # Choose the parameter set to use by index (e.g., 0 or 1)
 SELECTED_PARAMS=${PARAMS[$SLURM_ARRAY_TASK_ID]}