+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+- **a43f8cc:** Missed 'System' folder in previous commit
+- **b18cba5:** Created scripts for slurm based batch execution. Position calculation is done using the framerate and not actual time.
 - **3552754:** Added argument for setting LLM model
 - **0d490e2:** Modified npz plotter to have minimum value of 0 in y axis
 - **24be5fb:** Sorted command line arguments and added .bat file to run the script with various parameters.